# docs/openapi.yaml
openapi: 3.0.3
info:
  title: lensing API
  version: 0.1.0
  description: >-
    HTTP API of the lensing prediction lab: build feature datasets from a
    Qdrant corpus, launch and monitor training runs, promote runs into named
    models and predict against them, maintain the curated best-models group,
    manage model definitions (predictor + hyperparameter presets), author
    manual entries, and blend the test-split predictions of finished runs.
    Domain-specific vocabulary (entity noun, target noun, field names) comes
    from domain.toml — see GET /api/domain.
servers:
  - url: /api

# The API is unauthenticated (single-user lab server).
security: []
...
paths:
  /openapi.yaml:
    get:
      tags: [docs]
      summary: This OpenAPI spec
      description: Serves docs/openapi.yaml from disk so spec edits need no rebuild.
      responses:
        '200':
          description: OpenAPI document.
          content:
            application/yaml:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /datasets:
    get:
      tags: [datasets]
      summary: List datasets
      description: All dataset manifests, newest first. The PCA mean is stripped.
      responses:
        '200':
          description: Dataset manifests.
          content:
            application/json:
              schema:
                type: object
                properties:
                  datasets:
                    type: array
                    items: { $ref: '#/components/schemas/Manifest' }
    post:
      tags: [datasets]
      summary: Build a dataset
      description: >-
        Start an async dataset build (feature matrix + train/test split) from a
        Qdrant collection. Returns a build_id; poll GET /builds/{id}.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuildRequest' }
      responses:
        '200':
          description: Build accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  build_id: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
  /datasets/preflight:
    post:
      tags: [datasets]
      summary: Preview quality filtering
      description: >-
        Evaluate quality rules over the live corpus (payload-only scroll)
        without building anything: per-rule exclusion counts plus sample rows.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreflightRequest' }
      responses:
        '200':
          description: Per-rule counts, samples, and currency report.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                properties:
                  n_total: { type: integer }
                  n_excluded_total: { type: integer }
                  rules: { type: array, items: { type: object, additionalProperties: true } }
                  samples: {}
                  currency: {}
        '500':
          $ref: '#/components/responses/InternalError'
  /datasets/analyze:
    post:
      tags: [datasets]
      summary: Analyze feature spectrum and redundancy
      description: >-
        Async analyze job sharing the build prologue (scroll, currency,
        quality, split): full eigen-spectrum EVR preview plus feature
        redundancy report. Returns a job_id; poll GET /jobs/{id}.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuildRequest' }
      responses:
        '200':
          description: Analyze job accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
  /datasets/{id}:
    get:
      tags: [datasets]
      summary: Get a dataset manifest
      parameters: [{ $ref: '#/components/parameters/DatasetId' }]
      responses:
        '200':
          description: Dataset manifest (PCA mean stripped).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Manifest' }
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets/{id}/items:
    get:
      tags: [datasets]
      summary: Get dataset items
      description: The dataset's items.json (raw per-row metadata), as stored.
      parameters: [{ $ref: '#/components/parameters/DatasetId' }]
      responses:
        '200':
          description: items.json payload.
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets/{id}/split:
    get:
      tags: [datasets]
      summary: Get train/test split membership
      description: Train/test membership as row ids (not indices), for joining with items.json.
      parameters: [{ $ref: '#/components/parameters/DatasetId' }]
      responses:
        '200':
          description: Row-id membership.
          content:
            application/json:
              schema:
                type: object
                properties:
                  train: { type: array, items: { type: integer, format: int64 } }
                  test: { type: array, items: { type: integer, format: int64 } }
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets/{id}/archive:
    get:
      tags: [datasets]
      summary: Download a dataset as a tar.gz
      description: >-
        Tar.gz of the whole dataset directory, the transport remote training
        workers (`lensing-server worker --hub-url <this server>`) fetch
        datasets through.
      parameters: [{ $ref: '#/components/parameters/DatasetId' }]
      responses:
        '200':
          description: Gzipped tarball of the dataset directory.
          content:
            application/gzip:
              schema: { type: string, format: binary }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets/{id}/rename:
    post:
      tags: [datasets]
      summary: Rename a dataset
      description: >-
        Set a dataset's display name. The slug dataset_id (and directory) stay
        put, so run/definition lineage is unaffected.
      parameters: [{ $ref: '#/components/parameters/DatasetId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, description: "Display name; trimmed, 1..=80 chars, no control chars." }
      responses:
        '200':
          description: Updated manifest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Manifest' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /builds/{id}:
    get:
      tags: [builds]
      summary: Get build status
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Build status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BuildStatus' }
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/{id}:
    get:
      tags: [jobs]
      summary: Get async job status
      description: Status of a heavy async job (analyze, export).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Job status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobStatus' }
        '404':
          $ref: '#/components/responses/NotFound'
  /collections:
    get:
      tags: [collections]
      summary: List Qdrant collections
      description: Every Qdrant collection, for the build form's source picker.
      responses:
        '200':
          description: Collections plus the server's default source.
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections: { type: array, items: { type: string } }
                  source: { type: string }
        '500':
          $ref: '#/components/responses/InternalError'
  /collections/validate:
    post:
      tags: [collections]
      summary: Validate a collection's shape
      description: >-
        Probe a collection's vector config, filter match count, and payload
        field coverage against the domain. Shape problems are reported as data
        (always 200); only transport failures error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                collection: { type: string, description: Collection to probe; defaults to the server's. }
      responses:
        '200':
          description: Collection validation report.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '500':
          $ref: '#/components/responses/InternalError'
  /collections/export:
    post:
      tags: [collections]
      summary: Export a clean collection
      description: >-
        Write quality-filter survivors to a new Qdrant collection
        (<source>-clean-<suffix>); source untouched. Reconciled currency (and
        Convert-mode target values) are stamped into the exported payloads.
        Returns a job_id; poll GET /jobs/{id}.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportRequest' }
      responses:
        '200':
          description: Export job accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /runs:
    get:
      tags: [runs]
      summary: List runs
      description: >-
        All run metas, newest first: local run dirs plus database-backed rows
        (queued and remote-worker runs).
      responses:
        '200':
          description: Run metas.
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items: { $ref: '#/components/schemas/RunMeta' }
    post:
      tags: [runs]
      summary: Start a training run
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartRunRequest' }
      responses:
        '200':
          description: Run started (or enqueued when `queue` is set).
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
  /runs/{id}:
    get:
      tags: [runs]
      summary: Get a run meta
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: Run meta.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunMeta' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [runs]
      summary: Delete a finished run
      description: >-
        Delete a finished run (local directory or database-only row). Running
        runs must be stopped first; models promoted from the run keep working.
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /runs/{id}/predictions:
    get:
      tags: [runs]
      summary: Get run test-split predictions
      description: The run's predictions.json (test-split actual/predicted pairs).
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: Predictions.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Prediction' }
        '404':
          $ref: '#/components/responses/NotFound'
  /blend:
    post:
      tags: [blends]
      summary: Blend run predictions
      description: >-
        Blend the test-split predictions of two or more finished runs and
        report target-space metrics for the combination, each leg alone, and
        the pairwise residual correlations. Pure post-processing; nothing is
        trained or written. All runs must share the same test split.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BlendRequest' }
      responses:
        '200':
          description: Blend result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BlendResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /runs/{id}/events:
    get:
      tags: [runs]
      summary: Stream run progress (SSE)
      description: >-
        Server-Sent Events: replays recorded progress lines, then streams live
        events until the run is terminal. Each event's data is one JSON line
        (epoch / log / checkpoint / stopping / done / status). Queued and
        remote-worker runs replay database rows, polling while in flight.
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: Event stream.
          content:
            text/event-stream:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
  /runs/{id}/stop:
    post:
      tags: [runs]
      summary: Stop a running run
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
                  description: >-
                    false = graceful (STOP file; predictor finishes the epoch,
                    evaluates, checkpoints). true = stop-and-save with kill
                    escalation after a grace window.
      responses:
        '200':
          description: Stop acknowledged.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  mode: { type: string, enum: [graceful, force, kill] }
        '409':
          $ref: '#/components/responses/Conflict'
  /runs/{id}/viz:
    get:
      tags: [runs]
      summary: Get run architecture diagram
      description: Architecture SVG written by the predictor at training start (capability-gated).
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: SVG diagram.
          content:
            image/svg+xml:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
  /runs/{id}/blend:
    get:
      tags: [runs]
      summary: Get a blend run's blend record
      description: >-
        blend.json written by the blend predictor at the end of training:
        per-member kind/predictor/source, fitted weights, exclude_blocks and
        solo test metrics. Only blend runs have it; 404 otherwise.
      parameters: [{ $ref: '#/components/parameters/RunId' }]
      responses:
        '200':
          description: Blend record.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    DatasetId:
      name: id
      in: path
      required: true
      schema: { type: string }
    RunId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflicting state (e.g. name taken, run still running).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Internal error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    PreflightRequest:
      type: object
      properties:
        quality: { $ref: '#/components/schemas/QualityFilterConfig' }
        currency: { $ref: '#/components/schemas/CurrencyConfig' }
        sample: { type: integer, default: 8, description: Flagged samples per rule (capped at 50). }
        collection: { type: string, nullable: true }
...